requirements.txt
torch
torchvision
numpy
matplotlib
pandas
opencv-python
pydicom
pillow
pypng
tqdm
scikit-image